test(web): move the tailored-CV resolver test into web/tests
The resolver under test is web/ code, so its test belongs in web/tests/
where the web CI already collects it. Relocating it there lets the whole
core-side surface go: the test-all.mjs hook, the update-system.mjs
registration and the core-only-install skip guard all disappear, because
the core runner no longer needs to know the test exists.

Same assertions, same logic under test. The inline resolver hook stays:
web/src/lib/apply/cv.ts reaches career-ops.ts through the "@/" tsconfig
alias, which plain `node --test` does not resolve on its own.
