+++ test-all.mjs
@@ -15833,21 +15833,6 @@ console.log('\n60. Cover-letter template resolver (generate-cover-letter.mjs)');
   else fail('cover-resolver unit tests failed (run: node --test test/cover-resolver.test.mjs)');
 }
 
-console.log('\n60a. Tailored-CV resolver (web/src/lib/apply/cv.ts, PR #2156)');
-{
-  // web/ lives deliberately OUTSIDE the auto-updater's world (its own
-  // release-please component; see validate-system-paths-coverage.mjs
-  // EXCLUDE_PREFIXES) — a core-only install has no web/ tree, so warn (not
-  // fail) instead of crashing the suite.
-  if (existsSync(join(ROOT, 'web', 'src', 'lib', 'apply', 'cv.ts'))) {
-    const unit = run(NODE, ['--test', 'test/apply-cv-resolver.test.mjs']);
-    if (unit !== null) pass('apply-cv-resolver unit tests pass');
-    else fail('apply-cv-resolver unit tests failed (run: node --test test/apply-cv-resolver.test.mjs)');
-  } else {
-    warn('apply-cv-resolver tests skipped — web/ not present (core-only install)');
-  }
-}
-
 // ── 61. INTERVIEW-PREP URL ENTRY (#1816) ────────────────────────
 // Prompt-level slice: prep for a role that was never evaluated. Pins the
 // disambiguation rule (bare URL still routes to auto-pipeline), the
