+++ update-system.mjs
@@ -347,7 +347,6 @@ const SYSTEM_PATHS = [
   'test/cv-templates.test.mjs',
   'test/cover-resolver.test.mjs',
   'test/pipeline-lock.test.mjs',
-  'test/apply-cv-resolver.test.mjs',
   'test/profile-photo.test.mjs',
   'templates/cv-template.zh-minimal.html',
   'test/zh-minimal-template.test.mjs',
